+++ tests/Integration/ServiceAvailabilityTest.php
@@ -70,6 +70,11 @@ class ServiceAvailabilityTest extends IntegrationTestCase
             // Migration
             [\OxidEsales\WysiwygModule\Migration\Command\MigrateMediaUrlsToIdsCommand::class],
             [\OxidEsales\WysiwygModule\Migration\Command\MigrateMediaAltTextsCommand::class],
+            [\OxidEsales\WysiwygModule\Migration\Factory\MigrationReporterFactoryInterface::class],
+            [\OxidEsales\WysiwygModule\Migration\Repository\FieldMigrationRepositoryInterface::class],
+            [\OxidEsales\WysiwygModule\Migration\Service\FieldMigrationServiceInterface::class],
+            [\OxidEsales\WysiwygModule\Migration\Service\MediaMigrationResultFilterInterface::class],
+            [\OxidEsales\WysiwygModule\Migration\Service\MigrationServiceInterface::class],
 
             // Service
             [\OxidEsales\WysiwygModule\Service\EditorRendererInterface::class],